backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
django-cors-headers==4.9.0
django-filter==25.2
django-environ==0.12.0
psycopg2-binary==2.9.11
gunicorn==23.0.0
whitenoise==6.11.0
celery==5.5.3
redis==6.4.0
requests==2.32.5
phonenumbers==9.0.16
